Remove code which isn't strictly necessary
Much code in this service lives for the case where humans log in versus
how a system logs in.  A whole lot of code was removed in this light.
Possibly much more than the ideal case but this leaves us with less
untested code.  Though also in this limited set, the logout route was
not tested.

Utilities which provided only one function were moved into the main
file.